Move the `Power` members into `Battery` Closes #583
