+++ Xamarin.Essentials/Battery/Battery.android.cs
@@ -1,13 +1,51 @@
 ﻿using System;
 using Android.Content;
 using Android.OS;
-using Debug = System.Diagnostics.Debug;
 
 namespace Xamarin.Essentials
 {
     public static partial class Battery
     {
         static BatteryBroadcastReceiver batteryReceiver;
+        static EnergySaverBroadcastReceiver powerReceiver;
+
+        static void StartEnergySaverListeners()
+        {
+            if (!Platform.HasApiLevel(BuildVersionCodes.Lollipop))
+                return;
+
+            powerReceiver = new EnergySaverBroadcastReceiver(OnEnergySaverChanged);
+            Platform.AppContext.RegisterReceiver(powerReceiver, new IntentFilter(PowerManager.ActionPowerSaveModeChanged));
+        }
+
+        static void StopEnergySaverListeners()
+        {
+            if (!Platform.HasApiLevel(BuildVersionCodes.Lollipop))
+                return;
+
+            try
+            {
+                Platform.AppContext.UnregisterReceiver(powerReceiver);
+            }
+            catch (Java.Lang.IllegalArgumentException)
+            {
+                System.Diagnostics.Debug.WriteLine("Energy saver receiver already unregistered. Disposing of it.");
+            }
+            powerReceiver.Dispose();
+            powerReceiver = null;
+        }
+
+        static EnergySaverStatus PlatformEnergySaverStatus
+        {
+            get
+            {
+                var status = false;
+                if (Platform.HasApiLevel(BuildVersionCodes.Lollipop))
+                    status = Platform.PowerManager?.IsPowerSaveMode ?? false;
+
+                return status ? EnergySaverStatus.On : EnergySaverStatus.Off;
+            }
+        }
 
         static void StartBatteryListeners()
         {
@@ -25,7 +63,7 @@ namespace Xamarin.Essentials
             }
             catch (Java.Lang.IllegalArgumentException)
             {
-                Debug.WriteLine("Battery receiver already unregistered. Disposing of it.");
+                System.Diagnostics.Debug.WriteLine("Battery receiver already unregistered. Disposing of it.");
             }
             batteryReceiver.Dispose();
             batteryReceiver = null;
@@ -114,6 +152,22 @@ namespace Xamarin.Essentials
         }
 
         public BatteryBroadcastReceiver(Action onChanged) =>
+            this.onChanged = onChanged;
+
+        public override void OnReceive(Context context, Intent intent) =>
+            onChanged?.Invoke();
+    }
+
+    [BroadcastReceiver(Enabled = true, Exported = false, Label = "Essentials Energy Saver Broadcast Receiver")]
+    class EnergySaverBroadcastReceiver : BroadcastReceiver
+    {
+        Action onChanged;
+
+        public EnergySaverBroadcastReceiver()
+        {
+        }
+
+        public EnergySaverBroadcastReceiver(Action onChanged) =>
             this.onChanged = onChanged;
 
         public override void OnReceive(Context context, Intent intent) =>
