--- Xamarin.Essentials/Power/Power.netstandard.cs
@@ -1,14 +1,0 @@
-﻿namespace Xamarin.Essentials
-{
-    public static partial class Power
-    {
-        static void StartPowerListeners() =>
-            throw new NotImplementedInReferenceAssemblyException();
-
-        static void StopPowerListeners() =>
-            throw new NotImplementedInReferenceAssemblyException();
-
-        static EnergySaverStatus PlatformEnergySaverStatus =>
-            throw new NotImplementedInReferenceAssemblyException();
-    }
-}
