+++ docs/en/Xamarin.Essentials/AppInfo.xml
@@ -53,6 +53,24 @@
         <remarks></remarks>
       </Docs>
     </Member>
+    <Member MemberName="PackageName">
+      <MemberSignature Language="C#" Value="public static string PackageName { get; }" />
+      <MemberSignature Language="ILAsm" Value=".property string PackageName" />
+      <MemberSignature Language="DocId" Value="P:Xamarin.Essentials.AppInfo.PackageName" />
+      <MemberType>Property</MemberType>
+      <AssemblyInfo>
+        <AssemblyVersion>1.0.0.0</AssemblyVersion>
+        <AssemblyName>Xamarin.Essentials</AssemblyName>
+      </AssemblyInfo>
+      <ReturnValue>
+        <ReturnType>System.String</ReturnType>
+      </ReturnValue>
+      <Docs>
+        <summary>Gets the application package name or identifier.</summary>
+        <value>The package name or identifier.</value>
+        <remarks>On android and iOS, this is the application package name. On UWP, this is the application GUID.</remarks>
+      </Docs>
+    </Member>
     <Member MemberName="ShowSettingsUI">
       <MemberSignature Language="C#" Value="public static void ShowSettingsUI ();" />
       <MemberSignature Language="ILAsm" Value=".method public static hidebysig void ShowSettingsUI() cil managed" />
@@ -71,24 +89,6 @@
         <remarks>
           <para />
         </remarks>
-      </Docs>
-    </Member>
-    <Member MemberName="PackageName">
-      <MemberSignature Language="C#" Value="public static string PackageName { get; }" />
-      <MemberSignature Language="ILAsm" Value=".property string PackageName" />
-      <MemberSignature Language="DocId" Value="P:Xamarin.Essentials.AppInfo.PackageName" />
-      <MemberType>Property</MemberType>
-      <AssemblyInfo>
-        <AssemblyVersion>1.0.0.0</AssemblyVersion>
-        <AssemblyName>Xamarin.Essentials</AssemblyName>
-      </AssemblyInfo>
-      <ReturnValue>
-        <ReturnType>System.String</ReturnType>
-      </ReturnValue>
-      <Docs>
-        <summary>Gets the application package name or identifier.</summary>
-        <value>The package name or identifier.</value>
-        <remarks>On android and iOS, this is the application package name. On UWP, this is the application GUID.</remarks>
       </Docs>
     </Member>
     <Member MemberName="Version">
