+++ docs/en/Xamarin.Essentials/Battery.xml
@@ -62,6 +62,45 @@
         </remarks>
       </Docs>
     </Member>
+    <Member MemberName="EnergySaverStatus">
+      <MemberSignature Language="C#" Value="public static Xamarin.Essentials.EnergySaverStatus EnergySaverStatus { get; }" />
+      <MemberSignature Language="ILAsm" Value=".property valuetype Xamarin.Essentials.EnergySaverStatus EnergySaverStatus" />
+      <MemberSignature Language="DocId" Value="P:Xamarin.Essentials.Battery.EnergySaverStatus" />
+      <MemberType>Property</MemberType>
+      <AssemblyInfo>
+        <AssemblyName>Xamarin.Essentials</AssemblyName>
+        <AssemblyVersion>1.0.0.0</AssemblyVersion>
+      </AssemblyInfo>
+      <ReturnValue>
+        <ReturnType>Xamarin.Essentials.EnergySaverStatus</ReturnType>
+      </ReturnValue>
+      <Docs>
+        <summary>Gets the current energy saver status of the device.</summary>
+        <value>The current status of energy saver mode.</value>
+        <remarks>
+          <para></para>
+        </remarks>
+      </Docs>
+    </Member>
+    <Member MemberName="EnergySaverStatusChanged">
+      <MemberSignature Language="C#" Value="public static event EventHandler&lt;Xamarin.Essentials.EnergySaverStatusChangedEventArgs&gt; EnergySaverStatusChanged;" />
+      <MemberSignature Language="ILAsm" Value=".event class System.EventHandler`1&lt;class Xamarin.Essentials.EnergySaverStatusChangedEventArgs&gt; EnergySaverStatusChanged" />
+      <MemberSignature Language="DocId" Value="E:Xamarin.Essentials.Battery.EnergySaverStatusChanged" />
+      <MemberType>Event</MemberType>
+      <AssemblyInfo>
+        <AssemblyName>Xamarin.Essentials</AssemblyName>
+        <AssemblyVersion>1.0.0.0</AssemblyVersion>
+      </AssemblyInfo>
+      <ReturnValue>
+        <ReturnType>System.EventHandler&lt;Xamarin.Essentials.EnergySaverStatusChangedEventArgs&gt;</ReturnType>
+      </ReturnValue>
+      <Docs>
+        <summary>Event that is triggered when energy saver status changes.</summary>
+        <remarks>
+          <para></para>
+        </remarks>
+      </Docs>
+    </Member>
     <Member MemberName="PowerSource">
       <MemberSignature Language="C#" Value="public static Xamarin.Essentials.BatteryPowerSource PowerSource { get; }" />
       <MemberSignature Language="ILAsm" Value=".property valuetype Xamarin.Essentials.BatteryPowerSource PowerSource" />
