--- docs/en/Xamarin.Essentials/Power.xml
@@ -1,60 +1,0 @@
-<Type Name="Power" FullName="Xamarin.Essentials.Power">
-  <TypeSignature Language="C#" Value="public static class Power" />
-  <TypeSignature Language="ILAsm" Value=".class public auto ansi abstract sealed beforefieldinit Power extends System.Object" />
-  <TypeSignature Language="DocId" Value="T:Xamarin.Essentials.Power" />
-  <AssemblyInfo>
-    <AssemblyName>Xamarin.Essentials</AssemblyName>
-    <AssemblyVersion>1.0.0.0</AssemblyVersion>
-  </AssemblyInfo>
-  <Base>
-    <BaseTypeName>System.Object</BaseTypeName>
-  </Base>
-  <Interfaces />
-  <Docs>
-    <summary>Access to device specific power settings.</summary>
-    <remarks>
-      <para></para>
-    </remarks>
-  </Docs>
-  <Members>
-    <Member MemberName="EnergySaverStatus">
-      <MemberSignature Language="C#" Value="public static Xamarin.Essentials.EnergySaverStatus EnergySaverStatus { get; }" />
-      <MemberSignature Language="ILAsm" Value=".property valuetype Xamarin.Essentials.EnergySaverStatus EnergySaverStatus" />
-      <MemberSignature Language="DocId" Value="P:Xamarin.Essentials.Power.EnergySaverStatus" />
-      <MemberType>Property</MemberType>
-      <AssemblyInfo>
-        <AssemblyName>Xamarin.Essentials</AssemblyName>
-        <AssemblyVersion>1.0.0.0</AssemblyVersion>
-      </AssemblyInfo>
-      <ReturnValue>
-        <ReturnType>Xamarin.Essentials.EnergySaverStatus</ReturnType>
-      </ReturnValue>
-      <Docs>
-        <summary>Gets the current energy saver status of the device.</summary>
-        <value>The current status of energy saver mode.</value>
-        <remarks>
-          <para></para>
-        </remarks>
-      </Docs>
-    </Member>
-    <Member MemberName="EnergySaverStatusChanged">
-      <MemberSignature Language="C#" Value="public static event EventHandler&lt;Xamarin.Essentials.EnergySaverStatusChangedEventArgs&gt; EnergySaverStatusChanged;" />
-      <MemberSignature Language="ILAsm" Value=".event class System.EventHandler`1&lt;class Xamarin.Essentials.EnergySaverStatusChangedEventArgs&gt; EnergySaverStatusChanged" />
-      <MemberSignature Language="DocId" Value="E:Xamarin.Essentials.Power.EnergySaverStatusChanged" />
-      <MemberType>Event</MemberType>
-      <AssemblyInfo>
-        <AssemblyName>Xamarin.Essentials</AssemblyName>
-        <AssemblyVersion>1.0.0.0</AssemblyVersion>
-      </AssemblyInfo>
-      <ReturnValue>
-        <ReturnType>System.EventHandler&lt;Xamarin.Essentials.EnergySaverStatusChangedEventArgs&gt;</ReturnType>
-      </ReturnValue>
-      <Docs>
-        <summary>Event that is triggered when energy saver status changes.</summary>
-        <remarks>
-          <para></para>
-        </remarks>
-      </Docs>
-    </Member>
-  </Members>
-</Type>
